Django_flask_project/django_docker_nginx/requirements.txt
Django==3.1
psycopg2-binary
dj-static
gunicorn==20.0.4
asgiref==3.2.10
django-debug-toolbar==2.2
django-model-utils==4.0.0
elasticsearch==7.12.0
django-rest-elasticsearch